[sendreplydialog] Move focus on tab
Currently pressing tab inserts a tab character. To enable keyboard navigation it should however focus the next item in the focus chain (the send button). Given that this is a simple reply to a text message and not a fully fledged text editor there's probably no need to insert a tab character anyway

CCBUG: 441742